+++ baseKit/src/main/java/me/shetj/base/tools/debug/DebugFloatManager.kt
@@ -88,6 +88,9 @@ object DebugFloatManager : Application.ActivityLifecycleCallbacks {
                     lastTranslationX = x
                     lastTranslationY = y
                 }
+                onCloseListener = {
+                    isShow = false
+                }
             }
         }